# Improve performance by removing unnecessary locking.