### Contains
A performance optimization, archived by removing uncessary locking.

In my test this made the setBlocks method of WorldProvider more than twice
as fast.

The locking used to be done mainly by the main thread. Besides the main thread the light merging thread and the chunk mesh building thread were also locking. One reason why it probably got locked was the disposal of the chunk. When a chunk was no logner needed, it got made unusable which could have potentially be a problem for the background threads that were still using it. This patch makes the chunk remain readable even when the chunk is no longer needed. For the garbage collector it makes no difference and we can that way dispose chunks without having to let the main thread wait for background threads which were still reading the chunk(which should improve performance at that point too).

A possible small disadvantage of the removed locking is that if you do a mass block placement, that you could theoretically see very shortly a state where some of the new block placements are done while other aren't. However this is mainly theoretically and should not be a problem during normal gameplay. Even if it is visible in some mass block replacement scenarios, I think most would prefer that glitch over a lag.

The performance benchmarks have been done via the benchmark screen from PR #2454:

Without this patch:

benchmark: WorldProvider.setBlocks
iteration:  200 / 200
last duration:  113.7 ms
min duration:  64.2 ms
median duration:  121.5 ms
avg duration:  122.1 ms
max duration:  216.3 ms

With this patch:

benchmark: WorldProvider.setBlocks
iteration:  200 / 200
last duration:  59.0 ms
min duration:  28.9 ms
median duration:  56.2 ms
avg duration:  56.7 ms
max duration:  112.1 ms

### How to test

Merge #2454 first. Then run the benchmarks once with and without with patch. The setBlocks benchmarks can be run by first calling "showScreen benchmarkScreen"  however it is by default not selected so you need to change the benchmark selection first before you start it.

### Outstanding before merging
I would like to give other developers a chance to comment on this and to give feedback. So I think we should wait a few days before we merge this. @immortius and @skaldarnar might be interested in this PR.