@GooeyHub 2016-08-25
Hooray Jenkins reported success with all tests good!