@Cervator 2016-08-27
I tested out a merged version of this just to make sure the game works, a world looks right, and to get some numbers. While the `setBlock` option for the new benchmarks doesn't play along nicely on my system it looks like `setBlocks` does pretty well - way faster than `setBlock` and it manages to finish the 200 iterations.

Before this PR:

```
benchmark: WorldProvider.setBlocks
iteration:  200 / 200
last duration:  215.9 ms
min duration:  111.3 ms
median duration:  210.9 ms
avg duration:  215.3 ms
max duration:  371.8 ms
```

With this PR:

```
benchmark: WorldProvider.setBlocks
iteration:  200 / 200
last duration:  46.9 ms
min duration:  34.1 ms
median duration:  61.4 ms
avg duration:  63.0 ms
max duration:  208.5 ms
```

Both runs were with the standard 1 GB memory and various other things active on my system.

I don't think there's an API impact since the affected interfaces and such aren't exposed via `@API` - everything is internal. Just architecture. Nothing fails to compile :-)