@Cervator 2016-08-27
Wait ... `@API` in `package-info.java` :(

Back the label goes. In this case I take it we're changing API that just doesn't happen to have any external implementation. Why *is* `ChunkView` API anyway?