@flo 2016-08-28
@Cervator Wow, for you it is even more than 3 times faster.