@skaldarnar 2016-08-30
I can confirm the performance improvements for `setBlocks` as reported by you two. I'm fine with removing the locking alltogether from `ChunkView` - should we wait for a final word by @immortius? The performance improvements are worth some minor visual glitches from time to time imho. 

before patch:

	benchmark: WorldProvider.setBlocks
	iteration:  200 / 200
	last duration:  169,0 ms
	min duration:  67,5 ms
	median duration:  140,7 ms
	avg duration:  142,1 ms
	max duration:  207,3 ms

after patch: 

	benchmark: WorldProvider.setBlocks
	iteration:  200 / 200
	last duration:  56,5 ms
	min duration:  37,7 ms
	median duration:  49,6 ms
	avg duration:  51,2 ms
	max duration:  141,8 ms