@Cervator 2016-08-31
Maybe we can aim to merge it right after a soonish Alpha 4 release (couple days hopefully), then as one of the multiplayer tests on Saturday use the newest dev build with this merged to make sure we don't see anything weird on a server?