@Cervator 2016-09-09
Merging along with #2460 - while this is technically an API violation I'm not sure the affected code should be exposed in the first place. Might be more of an "undocumented" thing in classic API considerations (meant to be internal with only internal implementations known, but exposed anyway for whatever reason) rather than the occasional trivial violation we hop over while in Alpha.