@Cervator 2016-09-09
Oh by the way - @flo should we go ahead and add the benchmark screen to the shiny new developer tools menu?