@flo 2016-09-10
@Cervator theoretically yes, but it's not that trival to add properly as it use the Stone block from Core, which is why the screen is currently defined in the Core module which should probably not be accessed from a engine menu.