@Cervator 2016-09-10
Got it. Maybe another annotation-based thing to scan for, build the list of available benchmarks that way? But yeah not really as trivial as just connecting a button to a command.